// src/types/AskingPriceDataType.ts
interface AskingPriceDataTypes {
  askp1: String
  askp2: String
  askp3: String
  askp_rsqn1: String
  askp_rsqn2: String
  askp_rsqn3: String
  bidp1: String
  bidp2: String
  bidp3: String
  bidp_rsqn1: String
  bidp_rsqn2: String
  bidp_rsqn3: String
  stck_prpr: String
  total_askp_rsqn: String
  total_bidp_rsqn: String
}

export type { AskingPriceDataTypes }